Update analogin_api.c
Adding the function call: EADC_SetExtendSampleTime(eadc_base, chn, 0xF);  significantly improves the accuracy of the ADC reading. In this edit, the sample time is hard coded to 0xF. Ideally this would be a configurable parameter by the user. Perhaps there is a better place for this function to be added.